+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+
+### Added
+- `--videos-path` option to `npm start` and `npm run build-cache` scripts
+
 ## [1.1.0] - 2021-02-09
 
 ### Added